Move export tests to the new location.
